+++ CHANGELOG.md
@@ -12,6 +12,10 @@ The format is based on [keep a changelog](http://keepachangelog.com/) and this p
 - Add support for "count_multple" on "NakamaSocket.add_matchmaker_async()" and "NakamaSocket.add_matchmaker_party_async()"
 - Add C# support classes to better integrate the .NET client with the Mono version of Godot, allowing HTML5 exports to work
 
+### Fixed
+
+- Fix receiving "NakamaRTAPI.PartyClose" message
+
 ## [3.0.0] - 2022-03-28
 
 ### Added